refactored database management

+++ src/main/java/com/example/blanza/Expense.java
@@ -7,7 +7,7 @@ public class Expense extends FinancialEntity {
     private LocalDate date;
     private String paymentMethod;
 
-    public Expense(int current_user_id, String category, float amount, LocalDate time, String paymentMethod) {
+    public Expense(int current_user_id, String category, double amount, LocalDate time, String paymentMethod) {
         super(current_user_id, amount);
         this.category = category;
         this.amount = amount;
